+++ experiments/cassandra/src/main/scala/org/apache/gearpump/experiments/cassandra/CassandraStore.scala
@@ -25,7 +25,8 @@ import com.twitter.bijection.Bijection
 import org.apache.gearpump.TimeStamp
 import org.apache.gearpump.experiments.cassandra.AbstractCassandraStoreFactory._
 import org.apache.gearpump.experiments.cassandra.CassandraStore._
-import org.apache.gearpump.experiments.cassandra.lib.{CassandraConnector, CassandraConnectorConf, Logging, StoreConf}
+import org.apache.gearpump.experiments.cassandra.lib.connector.CassandraConnectorConf
+import org.apache.gearpump.experiments.cassandra.lib.{CassandraConnector, Logging, StoreConf}
 import org.apache.gearpump.streaming.transaction.api.{CheckpointStore, CheckpointStoreFactory}
 
 object AbstractCassandraStoreFactory {
