+++ experiments/cassandra/src/main/scala/org/apache/gearpump/experiments/cassandra/lib/connector/partitioner/BucketingRangeIndex.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gearpump.experiments.cassandra.lib.partitioner
+package org.apache.gearpump.experiments.cassandra.lib.connector.partitioner
 
 import scala.collection.mutable.ArrayBuffer
