+++ experiments/cassandra/src/main/scala/org/apache/gearpump/experiments/cassandra/lib/connector/partitioner/dht/TokenRange.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gearpump.experiments.cassandra.lib.partitioner.dht
+package org.apache.gearpump.experiments.cassandra.lib.connector.partitioner.dht
 
 import java.net.InetAddress
